backend/requirements/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements/requirements.in -o requirements/requirements.txt
amqp==5.3.1
    # via kombu
asgiref==3.8.1
    # via django
attrs==25.3.0
    # via
    #   jsonschema
    #   referencing
billiard==4.2.1
    # via celery
black==23.9.1
    # via -r requirements/requirements.in
celery==5.3.4
    # via -r requirements/requirements.in
click==8.2.1
    # via
    #   black
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.1
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.3.0
    # via celery
django==5.2.3
    # via
    #   django-filter
    #   djangorestframework
    #   drf-spectacular
    #   drf-spectacular-sidecar
    #   drf-yasg
django-environ==0.12.0
    # via -r requirements/requirements.in
django-filter==25.1
    # via -r requirements/requirements.in
djangorestframework==3.16.0
    # via
    #   -r requirements/requirements.in
    #   drf-spectacular
    #   drf-yasg
drf-spectacular==0.26.5
    # via -r requirements/requirements.in
drf-spectacular-sidecar==2025.6.1
    # via drf-spectacular
drf-yasg==1.21.10
    # via -r requirements/requirements.in
et-xmlfile==2.0.0
    # via openpyxl
flake8==7.1.1
    # via -r requirements/requirements.in
inflection==0.5.1
    # via
    #   drf-spectacular
    #   drf-yasg
isort==5.12.0
    # via -r requirements/requirements.in
jsonschema==4.24.0
    # via drf-spectacular
jsonschema-specifications==2025.4.1
    # via jsonschema
kombu==5.5.4
    # via celery
mccabe==0.7.0
    # via flake8
mypy-extensions==1.1.0
    # via black
numpy==1.26.0
    # via
    #   -r requirements/requirements.in
    #   pandas
openpyxl==3.0.10
    # via -r requirements/requirements.in
packaging==25.0
    # via
    #   black
    #   drf-yasg
    #   kombu
pandas==2.1.1
    # via -r requirements/requirements.in
pathspec==0.12.1
    # via black
platformdirs==4.3.8
    # via black
polars-lts-cpu==1.29.0
    # via -r requirements/requirements.in
prompt-toolkit==3.0.51
    # via click-repl
psycopg2==2.9.10
    # via -r requirements/requirements.in
pycodestyle==2.12.1
    # via flake8
pyflakes==3.2.0
    # via flake8
python-dateutil==2.9.0.post0
    # via
    #   celery
    #   pandas
pytz==2025.2
    # via
    #   drf-yasg
    #   pandas
pyyaml==6.0.2
    # via
    #   drf-spectacular
    #   drf-yasg
redis==5.0.1
    # via -r requirements/requirements.in
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
rpds-py==0.25.1
    # via
    #   jsonschema
    #   referencing
six==1.17.0
    # via python-dateutil
sqlparse==0.5.3
    # via django
typing-extensions==4.14.0
    # via referencing
tzdata==2025.2
    # via
    #   celery
    #   kombu
    #   pandas
uritemplate==4.2.0
    # via
    #   drf-spectacular
    #   drf-yasg
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.13
    # via prompt-toolkit
